+++ src/App.tsx
@@ -4,6 +4,8 @@ import { Dashboard } from './client/pages/Dashboard';
 import { OrdersPage } from './client/pages/orders/OrdersPage';
 import { OrderForm } from './client/pages/orders/OrderForm';
 import { OrderDetails } from './client/pages/orders/OrderDetails';
+import { BuyListPage } from './client/pages/orders/BuyListPage';
+import { BuyListStorePage } from './client/pages/orders/BuyListStorePage';
 import { StoresPage } from './client/pages/stores/StoresPage';
 import { StoreForm } from './client/pages/stores/StoreForm';
 import { StoreDetails } from './client/pages/stores/StoreDetails';
@@ -47,6 +49,8 @@ function App() {
               <Route path="/orders/new" element={<OrderForm />} />
               <Route path="/orders/:id" element={<OrderDetails />} />
               <Route path="/orders/:id/edit" element={<OrderForm />} />
+              <Route path="/buy-list" element={<BuyListPage />} />
+              <Route path="/buy-list/:storeId" element={<BuyListStorePage />} />
               <Route path="/stores" element={<StoresPage />} />
               <Route path="/stores/new" element={<StoreForm />} />
               <Route path="/stores/:id" element={<StoreDetails />} />